[DLMED] set data type for labels ≈

+++ examples/classification_3d/densenet_training_array.py
@@ -49,7 +49,7 @@ def main():
         "/workspace/data/medical/ixi/IXI-T1/IXI585-Guys-1130-T1.nii.gz",
     ]
     # 2 binary labels for gender classification: man and woman
-    labels = np.array([0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0])
+    labels = np.array([0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0], dtype=np.int64)
 
     # Define transforms
     train_transforms = Compose([ScaleIntensity(), AddChannel(), Resize((96, 96, 96)), RandRotate90(), ToTensor()])
